+++ packages/b2c-commerce-graphql-server/package.json
@@ -1,0 +1,46 @@
+{
+  "name": "b2c-graphql-server",
+  "version": "1.0.0",
+  "license": "ISC",
+  "private": true,
+  "dependencies": {
+    "aws-serverless-express": "^3.3.6",
+    "commerce-sdk": "^2.6.1",
+    "cors": "^2.8.5",
+    "express": "^4.17.2",
+    "express-graphql": "^0.12.0",
+    "graphql": "^16.2.0"
+  },
+  "devDependencies": {
+    "eslint": "7.31.0",
+    "eslint-config-prettier": "4.0.0",
+    "eslint-import-resolver-webpack": "0.10.0",
+    "eslint-plugin-header": "^3.1.1",
+    "eslint-plugin-import": "2.23.4",
+    "eslint-plugin-jsx-a11y": "6.4.1",
+    "eslint-plugin-prettier": "3.0.1",
+    "eslint-plugin-react": "7.24.0",
+    "prettier": "1.18.2"
+  },
+  "scripts": {
+    "format": "prettier --write \"**/*.{js,jsx}\"",
+    "lint": "eslint \"**/*.{js,jsx}\"",
+    "lint:fix": "npm run lint -- --fix",
+    "start": "node ssr.js",
+    "build": "rm -rf dist && mkdir dist && cp -R $(find . -depth 1 ! -iname 'node_modules' ! -iname 'dist' ! -iname '.*' ! -iname 'build.tar') dist/ && cd dist && npm install --production",
+    "push": "npm run build && sdk-upload --buildDirectory dist --projectSlug $npm_package_name"
+  },
+  "mobify": {
+      "ssrEnabled": true,
+      "ssrParameters": {
+          "ssrFunctionNodeVersion": "14.x"
+      },
+      "ssrOnly": [
+          "ssr.js",
+          "node_modules/**/*.*"
+      ],
+      "ssrShared": [
+          "intentionally-does-not-exist"
+      ]
+  }
+}
