--- packages/b2c-commerce-graphql/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "b2c-graphql",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "commerce-sdk": "^2.6.1",
-    "cors": "^2.8.5",
-    "express": "^4.17.2",
-    "express-graphql": "^0.12.0",
-    "graphql": "^16.2.0"
-  }
-}
